[Refactor] #2291 「配列等のテーブル類はファイル名の末尾をtableにする」という命名規則を外れてtypes になっていたmonster-attack-types.cpp/h をmonster-attack-table.cpp/h に改名した"
