+++ Bakabakaband/Bakabakaband/Bakabakaband.vcxproj
@@ -636,7 +636,7 @@
     <ClCompile Include="..\..\src\monster-attack\monster-attack-player.cpp" />
     <ClCompile Include="..\..\src\monster-attack\monster-attack-status.cpp" />
     <ClCompile Include="..\..\src\monster-attack\monster-attack-switcher.cpp" />
-    <ClCompile Include="..\..\src\monster-attack\monster-attack-types.cpp" />
+    <ClCompile Include="..\..\src\monster-attack\monster-attack-table.cpp" />
     <ClCompile Include="..\..\src\combat\attack-accuracy.cpp" />
     <ClCompile Include="..\..\src\monster-attack\monster-eating.cpp" />
     <ClCompile Include="..\..\src\player-attack\player-attack.cpp" />
@@ -1418,7 +1418,7 @@
     <ClInclude Include="..\..\src\monster-attack\monster-attack-player.h" />
     <ClInclude Include="..\..\src\monster-attack\monster-attack-status.h" />
     <ClInclude Include="..\..\src\monster-attack\monster-attack-switcher.h" />
-    <ClInclude Include="..\..\src\monster-attack\monster-attack-types.h" />
+    <ClInclude Include="..\..\src\monster-attack\monster-attack-table.h" />
     <ClInclude Include="..\..\src\combat\attack-accuracy.h" />
     <ClInclude Include="..\..\src\monster-attack\monster-eating.h" />
     <ClInclude Include="..\..\src\player-attack\player-attack-util.h" />
