+++ src/Makefile.am
@@ -486,7 +486,7 @@ hengband_SOURCES = \
 	monster-attack/monster-attack-processor.cpp monster-attack/monster-attack-processor.h \
 	monster-attack/monster-attack-status.cpp monster-attack/monster-attack-status.h \
 	monster-attack/monster-attack-switcher.cpp monster-attack/monster-attack-switcher.h \
-	monster-attack/monster-attack-types.cpp monster-attack/monster-attack-types.h \
+	monster-attack/monster-attack-table.cpp monster-attack/monster-attack-table.h \
 	monster-attack/monster-eating.cpp monster-attack/monster-eating.h \
 	\
 	monster-floor/monster-death.cpp monster-floor/monster-death.h \
