+++ src/lore/combat-types-setter.cpp
@@ -1,7 +1,7 @@
 ﻿#include "lore/combat-types-setter.h"
 #include "lore/lore-util.h"
 #include "monster-attack/monster-attack-effect.h"
-#include "monster-attack/monster-attack-types.h"
+#include "monster-attack/monster-attack-table.h"
 #include "system/monster-race-definition.h"
 #include "term/term-color-types.h"
