+++ src/melee/melee-util.h
@@ -2,7 +2,7 @@
 
 #include "effect/attribute-types.h"
 #include "monster-attack/monster-attack-effect.h"
-#include "monster-attack/monster-attack-types.h"
+#include "monster-attack/monster-attack-table.h"
 #include "system/angband.h"
 
 /* monster-attack-monster type*/
