+++ src/monster-attack/monster-attack-table.h
