+++ src/monster-floor/one-monster-placer.cpp
@@ -20,7 +20,7 @@
 #include "game-option/cheat-types.h"
 #include "grid/grid.h"
 #include "monster-attack/monster-attack-player.h"
-#include "monster-attack/monster-attack-types.h"
+#include "monster-attack/monster-attack-table.h"
 #include "monster-floor/monster-move.h"
 #include "monster-floor/monster-summon.h"
 #include "monster-floor/place-monster-types.h"
