+++ src/monster-race/monster-race-hook.cpp
@@ -1,7 +1,7 @@
 ﻿#include "monster-race/monster-race-hook.h"
 #include "dungeon/dungeon.h"
 #include "monster-attack/monster-attack-effect.h"
-#include "monster-attack/monster-attack-types.h"
+#include "monster-attack/monster-attack-table.h"
 #include "monster-race/monster-race.h"
 #include "monster-race/race-ability-mask.h"
 #include "monster-race/race-flags-resistance.h"
