+++ src/mspell/summon-checker.cpp
@@ -1,5 +1,5 @@
 ﻿#include "mspell/summon-checker.h"
-#include "monster-attack/monster-attack-types.h"
+#include "monster-attack/monster-attack-table.h"
 #include "monster-race/monster-race-hook.h"
 #include "monster-race/monster-race.h"
 #include "monster-race/race-flags1.h"
