+++ src/object/warning.cpp
@@ -12,7 +12,7 @@
 #include "grid/feature.h"
 #include "inventory/inventory-slot-types.h"
 #include "monster-attack/monster-attack-effect.h"
-#include "monster-attack/monster-attack-types.h"
+#include "monster-attack/monster-attack-table.h"
 #include "monster-race/monster-race.h"
 #include "monster-race/race-ability-flags.h"
 #include "monster-race/race-flags1.h"
