+++ src/view/display-lore-attacks.cpp
@@ -3,7 +3,7 @@
 #include "lore/combat-types-setter.h"
 #include "lore/lore-calculator.h"
 #include "lore/lore-util.h"
-#include "monster-attack/monster-attack-types.h"
+#include "monster-attack/monster-attack-table.h"
 #include "monster-race/race-flags1.h"
 #include "system/monster-race-definition.h"
 #include "term/term-color-types.h"
